+++ i18n/zh-cn.ts
@@ -63,6 +63,7 @@ const table: Partial<typeof __table> = {
     "techtree.notechtree": "没有科技树。",
     "techtree.techfolder": "科技树目录：",
     "worldmap.error.fieldnotindefaultmap": "default.map中缺少字段“{0}”。",
+    "worldmap.error.multiply256": "地图长宽必须为256的倍数：{0}x{1}。",
     "worldmap.error.provinceidtoolarge": "地块编号过大，最大值：{0}。",
     "worldmap.export.error": "无法导出地图：{0}。",
     "worldmap.export.error.imgformat": "无法导出地图：图像格式问题。",